# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/payments": {
      "post": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "initializePayment",
        "parameters": [
          {
            "name": "invoiceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/process": {
      "post": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "processPayment",
        "parameters": [
          {
            "name": "paymentId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/payments/{invoiceId}": {
      "get": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "getPaidPaymentByInvoiceId",
        "parameters": [
          {
            "name": "invoiceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Money": {
        "type": "object",
        "properties": {
          "price": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "fraction": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number"
          },
          "formatted": {
            "type": "string"
          }
        }
      },
      "Payment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "invoiceId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "PENDING",
              "PAID",
              "FAILED"
            ]
          },
          "amount": {
            "$ref": "#/components/schemas/Money"
          },
          "paymentProvider": {
            "type": "string",
            "enum": [
              "STRIPE",
              "PAYPAL"
            ]
          },
          "transactionReference": {
            "type": "string"
          },
          "cardSchema": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "dueTo": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}